+++ src/core/src/lib/utils.ts
@@ -2,6 +2,7 @@ import { FormlyFieldConfig } from './models';
 import { isObservable } from 'rxjs';
 import { AbstractControl } from '@angular/forms';
 import { FormlyFieldConfigCache } from './models';
+import { TemplateRef } from '@angular/core';
 
 export function getFieldId(formId: string, field: FormlyFieldConfig, index: string | number) {
   if (field.id) {
@@ -155,6 +156,7 @@ export function clone(value: any): any {
   if (
     !isObject(value) ||
     isObservable(value) ||
+    value instanceof TemplateRef ||
     /* instanceof SafeHtmlImpl */ value.changingThisBreaksApplicationSecurity ||
     ['RegExp', 'FileList', 'File', 'Blob'].indexOf(value.constructor.name) !== -1
   ) {