+++ CHANGELOG.md
@@ -93,6 +93,29 @@ To upgrade from 5.0 to 6.0 see [UPGRADE-6.0](./UPGRADE-6.0.md)
 * **core:** skip emitting modelChange on inputs changes ([414cda3](https://github.com/ngx-formly/ngx-formly/commit/414cda3f2c85526bc0a4ff02783972ca57fe1875))
 * **material:** removed double tag closing in datepicker type ([#2327](https://github.com/ngx-formly/ngx-formly/issues/2327)) ([153a596](https://github.com/ngx-formly/ngx-formly/commit/153a596915695225a881531b62a11abf505787e2))
 
+<a name="5.10.26"></a>
+## [5.10.26](https://github.com/ngx-formly/ngx-formly/compare/v5.10.25...v5.10.26) (2021-12-06)
+
+
+### Bug Fixes
+
+* **core:** improve detecting changes when using Function in expression ([#3080](https://github.com/ngx-formly/ngx-formly/issues/3080)) ([73551ee](https://github.com/ngx-formly/ngx-formly/commit/73551ee)), closes [#2993](https://github.com/ngx-formly/ngx-formly/issues/2993)
+* **json-schema:** take account of oneOf model change ([#3077](https://github.com/ngx-formly/ngx-formly/issues/3077)) ([1cbdc3e](https://github.com/ngx-formly/ngx-formly/commit/1cbdc3e)), closes [#3059](https://github.com/ngx-formly/ngx-formly/issues/3059)
+* **json-schema:** take account of undefined value for integer type ([#3078](https://github.com/ngx-formly/ngx-formly/issues/3078)) ([7bdd137](https://github.com/ngx-formly/ngx-formly/commit/7bdd137)), closes [#3075](https://github.com/ngx-formly/ngx-formly/issues/3075)
+
+
+
+<a name="5.10.25"></a>
+## [5.10.25](https://github.com/ngx-formly/ngx-formly/compare/v5.10.24...v5.10.25) (2021-12-04)
+
+
+### Bug Fixes
+
+* **core:** handle TemplateRef clone ([#3067](https://github.com/ngx-formly/ngx-formly/issues/3067)) ([8e1ba9f](https://github.com/ngx-formly/ngx-formly/commit/8e1ba9f)), closes [#2574](https://github.com/ngx-formly/ngx-formly/issues/2574)
+* **material:** assign prefix to templateOptions instead of field ([ecca5d3](https://github.com/ngx-formly/ngx-formly/commit/ecca5d3)), closes [#3074](https://github.com/ngx-formly/ngx-formly/issues/3074)
+
+
+
 <a name="5.10.24"></a>
 ## [5.10.24](https://github.com/ngx-formly/ngx-formly/compare/v5.10.23...v5.10.24) (2021-11-30)